Add `random` crate
The crate mimics the `RngCore` interface. It was named `random` to not
conflict with `rand`'s symbols.

I've also replaced locally defined RNGs in the `ixgbe` and
`ota_update_client` crates, that used the ... HPET timer count as
a seed.

A further development could be to expose an `add_entropy` function.
I'm not sure it's a major concren because a CSPRNG seeded with 32 bytes
of entropy should be plenty secure. However, if RDSEED or RDRAND don't
exist, it is absolutely necessary. Adding entropy would require locking
the CSPRNG, generating a random seed, combining it with the entropy, and
reseeding the CSPRNG. This would somehow need to be arbitrated to ensure
the the CSPRNG isn't being constantly reseeded, impacting performance.

Our QEMU configuration currently doesn't support RDSEED, so the backup
RDRAND is used.

Signed-off-by: Klim Tsoutsman <klim@tsoutsman.com>